LayoutTests/storage/indexeddb/objectStore-required-arguments.html

<!DOCTYPE html>
<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="resources/shared.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>

description("Test IndexedDB object store required arguments");

function test()
{
    removeVendorPrefixes();

    name = self.location.pathname;
    description = "My Test Database";
    request = evalAndLog("indexedDB.open(name, description)");
    request.onsuccess = openSuccess;
    request.onerror = unexpectedErrorCallback;
}

function openSuccess()
{
    db = evalAndLog("db = event.target.result");

    request = evalAndLog("request = db.setVersion('1')");
    request.onsuccess = createAndPopulateObjectStore;
    request.onerror = unexpectedErrorCallback;
}

function createAndPopulateObjectStore()
{
    deleteAllObjectStores(db);

    objectStore = evalAndLog("objectStore = db.createObjectStore('foo');");
    shouldThrow("objectStore.put();");
    shouldThrow("objectStore.add();");
    shouldThrow("objectStore.delete();");
    shouldThrow("objectStore.get();");
    shouldThrow("objectStore.createIndex();");
    shouldThrow("objectStore.createIndex('foo');");
    shouldThrow("objectStore.index();");
    shouldThrow("objectStore.deleteIndex();");
    finishJSTest();
}


test();

</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
